chore: rename global
